+++ routeros/resource_interface_bridge_vlan.go
@@ -51,6 +51,15 @@ func ResourceInterfaceBridgeVlan() *schema.Resource {
 		},
 		KeyDisabled: PropDisabledRw,
 		KeyDynamic:  PropDynamicRo,
+		"mvrp_forbidden": {
+			Type:     schema.TypeList,
+			Optional: true,
+			Elem: &schema.Schema{
+				Type: schema.TypeString,
+			},
+			Description:      "Ports that ignore all MRP messages and remains Not Registered (MT), as well as disables applicant from declaring specific VLAN ID (available since RouterOS 7.15).",
+			DiffSuppressFunc: AlwaysPresentNotUserProvided,
+		},
 		"tagged": {
 			Type:     schema.TypeList,
 			Optional: true,