+++ routeros/resource_interface_vlan.go
@@ -22,8 +22,17 @@ func ResourceInterfaceVlan() *schema.Resource {
 		KeyLoopProtectStatus:       PropLoopProtectStatusRo,
 		KeyMacAddress:              PropMacAddressRo,
 		KeyMtu:                     PropMtuRw(),
-		KeyName:                    PropNameForceNewRw,
-		KeyRunning:                 PropRunningRo,
+		"mvrp": {
+			Type:     schema.TypeBool,
+			Optional: true,
+			//Default:  false,
+			Description: "Specifies whether this VLAN should declare its attributes through Multiple VLAN Registration Protocol (MVRP) as an applicant (available since RouterOS 7.15). " +
+				"It can be used to register the VLAN with connected bridges that support MVRP. " +
+				"This property only has an effect when use-service-tag is disabled.",
+			DiffSuppressFunc: AlwaysPresentNotUserProvided,
+		},
+		KeyName:    PropNameForceNewRw,
+		KeyRunning: PropRunningRo,
 		"use_service_tag": {
 			Type:     schema.TypeBool,
 			Optional: true,